% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{tic_tac_toe}
\alias{tic_tac_toe}
\title{Play tic-tac-toe!}
\usage{
tic_tac_toe()
}
\description{
Invoke an interactive text-based tic-tac-toe game.
}
\details{
If you enter 'computer' as a player name, you can play against a bot.
}
